[test-integration] Minor testsuite improvements
