+++ test-integration/operator-tests/src/test/resources/infinispans/datagrid_service.yaml
@@ -1,7 +1,7 @@
 apiVersion: infinispan.org/v1
 kind: Infinispan
 metadata:
-  name: advanced-setup-a
+  name: datagrid-service
 spec:
   affinity:
     podAntiAffinity:
