+++ test-integration/pom.xml
@@ -11,7 +11,7 @@
 
     <properties>
         <!-- Dependencies versions -->
-        <version.infinispan>12.0.0.Final</version.infinispan>
+        <version.infinispan>12.1.1.Final</version.infinispan>
         <version.fabric8>4.13.0</version.fabric8>
         <version.junit>5.7.0</version.junit>
         <version.lombok>1.16.22</version.lombok>
